refactor: move ChunkIterator from core to filecoin-proofs
The ChunkIterator is a higher level primitive that is only used on the API
surface, hence moving it out of the core.
