+++ filecoin-proofs/Cargo.toml
@@ -35,6 +35,7 @@ fr32 = { path = "../fr32", version = "~7.0.0", default-features = false }
 once_cell = "1.8.0"
 blstrs = "0.7.0"
 ff = { version = "0.13.0", default-features = false }
+iowrap = "0.2.1"
 
 [dev-dependencies]
 criterion = "0.3"
