+++ storage-proofs-core/src/util.rs
@@ -1,7 +1,4 @@
-use std::{
-    cmp::min,
-    io::{self, Read},
-};
+use std::cmp::min;
 
 use anyhow::ensure;
 use bellperson::{
@@ -9,15 +6,11 @@ use bellperson::{
     ConstraintSystem, SynthesisError,
 };
 use ff::PrimeField;
-use iowrap::ReadMany;
 use merkletree::merkle::get_merkle_tree_row_count;
 
 use crate::{error::Error, settings::SETTINGS};
 
 pub const NODE_SIZE: usize = 32;
-
-/// The number of bytes that are read from the reader at once.
-const READER_CHUNK_SIZE: usize = 4096;
 
 /// Returns the start position of the data, 0-indexed.
 pub fn data_at_node_offset(v: usize) -> usize {
@@ -181,46 +174,6 @@ pub fn default_rows_to_discard(leafs: usize, arity: usize) -> usize {
         2 => min(max_rows_to_discard, 7),
         4 => min(max_rows_to_discard, 5),
         _ => min(max_rows_to_discard, rows_to_discard),
-    }
-}
-
-// Based on
-// https://stackoverflow.com/questions/73145503/iterator-for-reading-file-chunks/73145594#73145594
-/// Chunks the given reader to the given size.
-///
-/// If the end is reached and there are a few bytes left, that don't fill a full chunk, those bytes
-/// are returned.
-pub struct ChunkIterator<R> {
-    reader: R,
-    chunk_size: usize,
-}
-
-impl<R: Read> ChunkIterator<R> {
-    /// Return a new iterator with a default chunk size of 4KiB.
-    pub fn new(reader: R) -> Self {
-        Self {
-            reader,
-            chunk_size: READER_CHUNK_SIZE,
-        }
-    }
-
-    pub const fn chunk_size(&self) -> usize {
-        self.chunk_size
-    }
-}
-
-impl<R: Read> Iterator for ChunkIterator<R> {
-    type Item = io::Result<Vec<u8>>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let mut buffer = vec![0u8; self.chunk_size];
-        match self.reader.read_many(&mut buffer) {
-            Ok(bytes_read) if bytes_read == self.chunk_size => Some(Ok(buffer)),
-            // A position of 0 indicates end of file.
-            Ok(bytes_read) if bytes_read == 0 => None,
-            Ok(bytes_read) => Some(Ok(buffer[..bytes_read].to_vec())),
-            Err(error) => Some(Err(error)),
-        }
     }
 }
